Billboard & Association Rule Mining Solutions
Got an output to the Association Rule Mining question that looked similar to what we saw in class. Decided to do it in python rather than R, so would love to see what yall think of what I did. Tried to explain what my code did in the comments and towards the bottom in my question conclusion.